# docs/swagger.yaml
definitions:
...
  CreatePostRequest:
    description: Request payload for creating a new post
    properties:
      content:
        type: string
      idempotency_key:
        type: string
      title:
        type: string
    type: object
  CreatePostResponse:
    description: Response with ID of the created post
    properties:
      post_id:
        type: string
    type: object
...
  PostResponse:
    properties:
      author:
        $ref: '#/definitions/UserResponse'
      content:
        type: string
      created_at:
        type: string
      post_id:
        type: string
      status:
        $ref: '#/definitions/TypePostStatus'
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  TypePostStatus:
    enum:
    - draft
    - published
    type: string
    x-enum-varnames:
    - DraftStatus
    - PublishedStatus
...
  UserResponse:
    properties:
      email:
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /posts:
    get:
      consumes:
      - application/json
      description: Read all posts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/PostResponse'
            type: array
        "400":
          description: Incorrect body\nRefresh token expired or incorrect
        "403":
          description: Access denied
        "404":
          description: Post not found
      security:
      - BearerAuth: []
      summary: Read post
      tags:
      - Reader
    post:
      consumes:
      - application/json
      description: Create new post
      parameters:
      - description: Create post data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CreatePostResponse'
        "400":
          description: Incorrect body
        "403":
          description: Incorrect user
        "409":
          description: Idempotency key already used
      security:
      - BearerAuth: []
      summary: Create post
      tags:
      - Poster
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {jwt_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"